git-svn-id: http://svn.seqan.de/seqan/trunk@2319 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/tests/refinement/test_graph_interval_tree.h
@@ -304,7 +304,7 @@ void Test_GraphIntervalTree()
 	IntervalTreeTest<int,RandomCenter,StorePointsOnly>();
 
 //	std::cout << "\nMidCenter ";
-	IntervalTreeTest<double,MidCenter,StoreIntervals>();
+	IntervalTreeTest<unsigned,MidCenter,StoreIntervals>();
 	
 //	std::cout << "\nComputeCenter ";
 	IntervalTreeTest<int,ComputeCenter,StoreIntervals>();
